% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{assertClass}
\alias{assertClass}
\title{Stop if an object is not from the good class and print an error message.}
\usage{
assertClass(object, className, message)
}
\arguments{
\item{object}{An object which class needs to be checked.}

\item{className}{A string indicating the class to check.}

\item{message}{An error message.}
}
\description{
Stop if an object is not from the good class and print an error message.
}
\keyword{internal}
