% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{assertExportFormat}
\alias{assertExportFormat}
\title{Stop if the data to export is in a wrong format.}
\usage{
assertExportFormat(data)
}
\arguments{
\item{data}{An imExport object of the good format.}
}
\description{
Stop if the data to export is in a wrong format.
}
\keyword{internal}
