% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{assertUploadFormat}
\alias{assertUploadFormat}
\title{Stop if the data to upload is in a wrong format.}
\usage{
assertUploadFormat(data)
}
\arguments{
\item{data}{An imSignals or imMetrics object of the good format.}
}
\description{
Stop if the data to upload is in a wrong format.
}
\keyword{internal}
