% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{assertValid}
\alias{assertValid}
\title{Stop if a condition is not fulfilled and print an error message.}
\usage{
assertValid(condition, message)
}
\arguments{
\item{condition}{R expressions which should evaluate to TRUE.}

\item{message}{An error message.}
}
\description{
Stop if a condition is not fulfilled and print an error message.
}
\keyword{internal}
