% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{createExport}
\alias{createExport}
\title{Create an export file at a specific location and append metadata to it if provided. Note that a column with the
study name will be appended to each export.}
\usage{
createExport(study, data, outputDirectory, fileName, metadata = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{data}{A data.table containing the export metrics to save.}

\item{outputDirectory}{The path where the file should be created.}

\item{fileName}{The name of the file to create (should finish with .csv).}

\item{metadata}{Optional - a data.table with metadata information. Column names will be converted to metadata headers
and there must be a row corresponding to each data column.}
}
\description{
Create an export file at a specific location and append metadata to it if provided. Note that a column with the
study name will be appended to each export.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
data <- data.frame("Respondent Name" = "Test", "Metric1" = seq(1:100), "Metric2" = rep(0, 100))
createExport(study, data, outputDirectory = "C:/Documents", fileName = "textExport.csv")

# Adding some metadata to the data
metadata <- data.table("Units" = c("", "ms", ""), "Description" = c("Desc1", "Desc2", "Desc"))
createExport(study, data, outputDirectory = "C:/Documents", fileName = "textExport.csv", metadata)
}
}
