% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{createImObject}
\alias{createImObject}
\title{Create an iMotions object.}
\usage{
createImObject(data, typeOfObject)
}
\arguments{
\item{data}{a list/data.frame that needs to be changed into an iMotions object.}

\item{typeOfObject}{a string that specifies the type of object wanted ("Stimulus","Respondent", "Segment", "Sensor",
"Interval").}
}
\value{
An iMotions object of the specified type (data.table).
}
\description{
Based on a data.table and the type of object wanted, create an imObject or imObjectList.
}
\examples{
\dontrun{
object <- list(name = "example")
respondentsObject <- imotionsApi:::createImObject(object, "Respondent")
stimuliObject <- imotionsApi:::createImObject(object, "Stimulus")
segmentsObject <- imotionsApi:::createImObject(object, "Segment")
sensorsObject <- imotionsApi:::createImObject(object, "Sensor")
}
}
\keyword{internal}
