% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAoiDetailsUrl.imAOI}
\alias{getAoiDetailsUrl.imAOI}
\title{getAoiDetailsUrl.imAOI method to return the path/url to a specific AOI's details information.}
\usage{
\method{getAoiDetailsUrl}{imAOI}(study, imObject, respondent = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imAOI or imStimulus object of interest.}

\item{respondent}{Optional - An imRespondent object as returned from  \code{\link{getRespondents}}.}
}
\description{
getAoiDetailsUrl.imAOI method to return the path/url to a specific AOI's details information.
}
\keyword{internal}
