% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAoiDetailsUrl.imStimulus}
\alias{getAoiDetailsUrl.imStimulus}
\title{getAoiDetailsUrl.imStimulus method to return the path/url to all AOIs' details for a specific stimulus.}
\usage{
\method{getAoiDetailsUrl}{imStimulus}(study, imObject, respondent = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imAOI or imStimulus object of interest.}

\item{respondent}{Optional - An imRespondent object as returned from  \code{\link{getRespondents}}.}
}
\description{
getAoiDetailsUrl.imStimulus method to return the path/url to all AOIs' details for a specific stimulus.
}
\keyword{internal}
