% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAoiMetricsUrl.imRespondent}
\alias{getAoiMetricsUrl.imRespondent}
\title{getAoiMetricsUrl method to return the path/url to retrieve AOI metrics for a respondent.}
\usage{
\method{getAoiMetricsUrl}{imRespondent}(study, imObject, AOI)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent object of interest.}

\item{AOI}{An imAOI object as returned from \code{\link{getAois}}.}
}
\description{
getAoiMetricsUrl method to return the path/url to retrieve AOI metrics for a respondent.
}
\keyword{internal}
