% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAoiMetricsUrl}
\alias{getAoiMetricsUrl}
\title{Generic getAoiMetricsUrl function that takes as parameter a study, respondent object and AOI.}
\usage{
getAoiMetricsUrl(study, imObject, AOI)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent object of interest.}

\item{AOI}{An imAOI object as returned from \code{\link{getAois}}.}
}
\description{
Return the path/url to retrieve AOI metrics for the given object.
}
\keyword{internal}
