% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAoiRespondentData}
\alias{getAoiRespondentData}
\title{Get the inOutGaze information, inOutMouse information and AOI's intervals for a specific AOI/respondent
combination. Note that imAOI object, by definition, are linked to a specific stimulus.}
\usage{
getAoiRespondentData(study, AOI, respondent)
}
\arguments{
\item{study}{An imStudy object as returned from imStudy()}

\item{AOI}{An imAOI object as returned from \code{\link{getAois}}.}

\item{respondent}{An imRespondent object as returned from \code{\link{getRespondents}}.}
}
\value{
A list with inOutGaze/inOutMouse information for the specific AOI/respondent combination and an
        imIntervalList object (data.table) composed of the start, end, duration, id and name of this AOI.
}
\description{
The inOutGaze data.table has a IsGazeInAOI column that is TRUE when a gaze was recorded inside the AOI and FALSE if
outside (timestamps correspond to the actual gazepoint Timestamp). To reduce the size of the file created,
only timestamps where a change of value occur are given. If the AOI was never active, the table is empty.
}
\details{
The inOutMouse data.table has a IsMouseInAOI column that is TRUE when a click was recorded inside the AOI and
FALSE if outside (timestamps correspond to the actual Timestamp of each click). If no click was recorded or if the
AOI was never active, the table is empty.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
AOIs <- imotionsApi::getAois(study)
respondents <- imotionsApi::getRespondents(study, AOI = AOIs[1, ])
AOIData <- imotionsApi::getAoiRespondentData(study, AOIs[1, ], respondents[1, ])

# Retrieving list items
inOutData <- AOIData$inOutData
intervals <- AOIData$intervals
}
}
