% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAoiRespondentTouchData}
\alias{getAoiRespondentTouchData}
\title{Get the contact in/out data for a specific (touch actor, AOI)/respondent combination.}
\usage{
getAoiRespondentTouchData(study, touchAoi, respondent)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{touchAoi}{One row of \code{\link{getTouchActorAois}}, i.e. an AOI carrying a contact in/out fileId.}

\item{respondent}{An imRespondent object as returned from \code{\link{getRespondents}}.}
}
\value{
A list with an inOutContact data.table (change-only rows of Timestamp, IsContactInAOI, HandType, HandId)
        and an imIntervalList object with the AOI's active fragments. NULL if no in/out file was found.
}
\description{
Get the contact in/out data for a specific (touch actor, AOI)/respondent combination.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
stimulus <- imotionsApi::getStimuli(study)[1, ]
respondent <- imotionsApi::getRespondents(study, stimulus = stimulus)[1, ]
touchAois <- imotionsApi::getTouchActorAois(study, stimulus, respondent)
contactData <- imotionsApi::getAoiRespondentTouchData(study, touchAois[1, ], respondent)

# Retrieving list items
inOutContact <- contactData$inOutContact
intervals <- contactData$intervals
}
}
