% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAoiRespondentTouchMetrics}
\alias{getAoiRespondentTouchMetrics}
\title{Get the contact metrics for a specific (touch actor, AOI)/respondent combination.}
\usage{
getAoiRespondentTouchMetrics(study, touchAoi, respondent)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{touchAoi}{One row of \code{\link{getTouchActorAois}}, i.e. an AOI carrying a contact in/out fileId.}

\item{respondent}{An imRespondent object as returned from \code{\link{getRespondents}}.}
}
\value{
A data.table of one row (imMetrics object) with the contact metrics, or NULL if none were computed yet.
}
\description{
Get the contact metrics for a specific (touch actor, AOI)/respondent combination.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
stimulus <- imotionsApi::getStimuli(study)[1, ]
respondent <- imotionsApi::getRespondents(study, stimulus = stimulus)[1, ]
touchAois <- imotionsApi::getTouchActorAois(study, stimulus, respondent)
metrics <- imotionsApi::getAoiRespondentTouchMetrics(study, touchAois[1, ], respondent)
}
}
