% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAois}
\alias{getAois}
\title{Get AOIs from a study.}
\usage{
getAois(
  study,
  stimulus = NULL,
  respondent = NULL,
  generateInOutFiles = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{stimulus}{Optional - An imStimulus object as returned from \code{\link{getStimuli}}.}

\item{respondent}{Optional - An imRespondent object as returned from \code{\link{getRespondents}}.}

\item{generateInOutFiles}{Optional - A boolean indicating whether the corresponding InOut files should be generated
and linked to each imAOI object.}

\item{verbose}{Optional - A boolean indicating whether warnings from privateAOIfiltering should be shown.}
}
\value{
An imAOIList object (data.table) with all AOIs of interest.
}
\description{
Generic getAois function that takes as parameter a study object and optionally a respondent/stimulus object.
In case no AOIs is defined for the combination, return NULL.
}
\details{
Important to note: to speed up the computation of gazes falling in/out AOIs at the respondent/stimulus level, an
optional parameter generateInOutFiles can be enabled (both stimulus AND respondent arguments need to be provided).
When enabled, a file will be generated for each AOI of this stimulus/respondent combination containing information
regarding AOI activation/deactivation and gazes/clicks falling in. Filepaths to these newly generated files
will be stored in their corresponding AOI object. If available, these filepaths will then directly be used by the
\code{\link{getAoiRespondentData}} function instead of re-generating the files. This is particularly useful in case
multiple AOIs are defined for the same stimulus. For remote studies, InOutFiles are already generated so this
parameter doesn't apply.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
stimuli <- imotionsApi::getStimuli(study)
respondents <- imotionsApi::getRespondents(study)

## Get all AOIs in the study
AOIs <- imotionsApi::getAois(study)

## Get all AOIs defined for a specific stimulus
AOIs <- imotionsApi::getAois(study, stimulus = stimuli[1, ])

## Get all AOIs defined for a specific respondent
AOIs <- imotionsApi::getAois(study, respondent = respondents[1, ])

## Get all AOIs defined for a specific stimulus/respondent combination
AOIs <- imotionsApi::getAois(study, respondent = respondents[1, ], stimulus = stimuli[1, ])

## Get all AOIs defined for a specific stimulus/respondent combination and process their InOut data
AOIs <- imotionsApi::getAois(study, respondent = respondents[1, ], stimulus = stimuli[1, ],
                             generateInOutFiles = T)

print(AOIs$fileId) # a field "fileId" should have been added with the path to the InOut file
}
}
