% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAoisUrl}
\alias{getAoisUrl}
\title{getAoiUrl function that takes as parameter a study object and optionally a respondent or a stimulus id.}
\usage{
getAoisUrl(study, stimulusId = NULL, respondentId = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{stimulusId}{Optional - the id of the stimulus of interest.}

\item{respondentId}{Optional - the id of the respondent of interest.}
}
\description{
Return the path/url to the AOIs for this study/respondent/stimulus.
}
\keyword{internal}
