% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getJSON}
\alias{getJSON}
\title{Retrieve a JSON file from the indicated path/url - optionally allow the user to add more information to know when an
error occurred.}
\usage{
getJSON(connection, url, message = NULL, auth = TRUE, ...)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}

\item{url}{The url/path where the JSON file is located.}

\item{message}{Optional - a short message indicating which steps are getting performed to get a more indicative
error message.}

\item{auth}{Whether to add the bearer token to the request, by default it is added to the request.}

\item{...}{Optional - arguments passed to jsonlite::fromJSON.}
}
\value{
The retrieved JSON file.
}
\description{
Retrieve a JSON file from the indicated path/url - optionally allow the user to add more information to know when an
error occurred.
}
\keyword{internal}
