% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getParentStimulusNames}
\alias{getParentStimulusNames}
\title{Get parent stimulus names from a vector of parent ids and a stimuli data.table.}
\usage{
getParentStimulusNames(parentIds, stimuli)
}
\arguments{
\item{parentIds}{A vector of parent stimulus ids.}

\item{stimuli}{An imStimulusList object as returned from \code{\link{getStimuli}}.}
}
\value{
The parent stimulus names corresponding to the given ids.
}
\description{
The parent name is set to "" if a stimulus has no parent and to "N/A" if the stimulus parent was not found in the
stimuli table (e.g. if the parent stimulus was not included in an analysis).
}
\keyword{internal}
