% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getRespondent}
\alias{getRespondent}
\title{Get a specific respondent from a study.}
\usage{
getRespondent(study, respondentId)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{respondentId}{The id of the respondent you would like to retrieve.}
}
\value{
An imRespondent object (data.table) containing the respondent of interest.
}
\description{
Available respondents can be found with \code{\link{getRespondents}}.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)
respondent <- imotionsApi::getRespondent(study, respondents$id[1])
}
}
