% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getRespondentIntervals}
\alias{getRespondentIntervals}
\title{Get the list of time intervals (imIntervalList) for a given respondent.}
\usage{
getRespondentIntervals(
  study,
  respondent,
  type = c("Stimulus", "Scene", "Annotation")
)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{respondent}{An imRespondent object as returned from \code{\link{getRespondents}}.}

\item{type}{The type of intervals to retrieve (can be set to Stimulus, Scene and/or Annotation).}
}
\value{
An imIntervalList object (data.table) composed of the start, end, duration, parent stimulus, id and name of
        each stimulus/scene/annotation. Annotations comments are also included.
}
\description{
This imIntervalList is composed of stimuli/scenes/annotations intervals.
For remote study, only stimuli intervals are supported as of now.
Note that AOIs intervals can be retrieved using \code{\link{getAoiRespondentData}}.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)
intervals <- imotionsApi::getRespondentIntervals(study, respondents[1, ])

# Get only the stimuli intervals
intervals <- imotionsApi::getRespondentIntervals(study, respondents[1, ], type = "Stimulus")
}
}
