% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getRespondents}
\alias{getRespondents}
\title{Get respondents from a study.}
\usage{
getRespondents(
  study,
  stimulus = NULL,
  AOI = NULL,
  segment = NULL,
  keepRespondentVariables = FALSE
)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{stimulus}{Optional - An imStimulus object as returned from \code{\link{getStimuli}}.}

\item{AOI}{Optional - An imAOI object as returned from \code{\link{getAois}}.}

\item{segment}{Optional - An imSegment object as returned from \code{\link{getSegments}}.}

\item{keepRespondentVariables}{Optional - A boolean indicating whether respondent variables should be kept (if they
are available). If this has the default value of FALSE, only the "group" variable is
exposed.}
}
\value{
An imRespondentList object (data.table) with all respondents of interest.
}
\description{
Generic getRespondents function that takes as parameter a study object and optionally a stimulus/AOI/segment
object. As AOIs are linked to a stimulus, it is not possible to provide both of them.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
stimuli <- imotionsApi::getStimuli(study)
segments <- imotionsApi::getSegments(study)
AOIs <- imotionsApi::getAois(study)

## Get all respondents in the study
respondents <- imotionsApi::getRespondents(study)

## Get all respondents exposed to a specific stimulus
respondents <- imotionsApi::getRespondents(study, stimulus = stimuli[1, ])

## Get all respondents in a specific segment
respondents <- imotionsApi::getRespondents(study, segment = segments[1, ])

## Get all respondents for whom a specific AOI has been defined
respondents <- imotionsApi::getRespondents(study, AOI = AOIs[1, ])

## Get all respondents in a specific segment exposed to a specific stimulus
respondents <- imotionsApi::getRespondents(study, stimulus = stimuli[1, ], segment = segments[1, ])

## Get all respondents in a specific segment for whom a specific AOI has been defined
respondents <- imotionsApi::getRespondents(study, AOI = AOIs[1, ], segment = segments[1, ])

## Get all respondents in the study and access their available variables
respondents <- imotionsApi::getRespondents(study, keepRespondentVariables = TRUE)
}
}
