% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getRespondentScenesUrl}
\alias{getRespondentScenesUrl}
\title{Return the path/url to the scenes of a specific respondent.}
\usage{
getRespondentScenesUrl(study, respondent)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{respondent}{An imRespondent object as returned from \code{\link{getRespondents}}.}
}
\description{
Return the path/url to the scenes of a specific respondent.
}
\keyword{internal}
