% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getRespondentSensors}
\alias{getRespondentSensors}
\title{Get all sensors available for a given respondent.}
\usage{
getRespondentSensors(study, respondent, stimulus = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{respondent}{An imRespondent object as returned from \code{\link{getRespondents}}.}

\item{stimulus}{Optional - an imStimulus object as returned from \code{\link{getStimuli}} to retrieve sensors
specific to this stimulus.}
}
\value{
An imSensorList object (data.table) with all sensors collected for the respondent of interest.
}
\description{
Retrieves detailed information about sensors available for a given respondent.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)
sensors <- imotionsApi::getRespondentSensors(study, respondents[1, ])

# Get sensors for a specific stimulus
stimuli <- imotionsApi::getStimuli(study)
sensors <- imotionsApi::getRespondentSensors(study, respondents[1, ], stimuli[1, ])
}
}
