% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getSegment}
\alias{getSegment}
\title{Get a specific segment from a study.}
\usage{
getSegment(study, segmentId)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{segmentId}{The id of the segment you would like to retrieve.}
}
\value{
An imSegment object (data.table) containing the segment of interest.
}
\description{
Available segments can be found with \code{\link{getSegments}}.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
segments <- imotionsApi::getSegments(study)
segment <- imotionsApi::getSegment(study, segments$id[1])
}
}
