% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getSegments}
\alias{getSegments}
\title{Get all segments from a study.}
\usage{
getSegments(study)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}
}
\value{
An imSegmentList object (data.table) containing all segments from the study.
}
\description{
Retrieves detailed information about segments in the study.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
segments <- imotionsApi::getSegments(study)
}
}
