% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getSensorData}
\alias{getSensorData}
\title{Download data corresponding to a specific sensor (signals/metrics).}
\usage{
getSensorData(study, sensor, signalsName = NULL, intervals = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{sensor}{An imSensor object as returned from \code{\link{getRespondentSensors}}.}

\item{signalsName}{Optional - A vector of specific signals name you would like to return.}

\item{intervals}{Optional - An imInterval or imIntervalList object with start/end of data to subset as given by
\code{\link{getRespondentIntervals}}.}
}
\value{
An imData object (data.table) containing the signals (imSignals) or metrics (imMetrics) for the sensor of
        interest.
}
\description{
Available sensors in your study can be listed using the \code{\link{getRespondentSensors}}.
}
\details{
Signals always have a "Timestamp" column and are unique to a given respondent and a given sensor source.
Metrics are stored as a special sensor, also specific to a given respondent.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)
sensors <- imotionsApi::getRespondentSensors(study, respondents[1, ])
data <- imotionsApi::getSensorData(study, sensors[1, ])
}
}
