% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getSensorDataUrl}
\alias{getSensorDataUrl}
\title{Return the path/url to the signals of a specific sensor.}
\usage{
getSensorDataUrl(study, sensor)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{sensor}{An imSensor object as returned from \code{\link{getSensors}}.}
}
\description{
Return the path/url to the signals of a specific sensor.
}
\keyword{internal}
