% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getSensorMetaData}
\alias{getSensorMetaData}
\title{Download metadata corresponding to a specific sensor (sensor and signals metadata).}
\usage{
getSensorMetaData(sensor)
}
\arguments{
\item{sensor}{An imSensor object as returned from \code{\link{getRespondentSensors}}.}
}
\value{
A list of two data frames containing respectively the sensor and the signals metadata.
}
\description{
Available sensors in your study can be listed using the \code{\link{getRespondentSensors}}.
}
\details{
Metadata contains sensor metadata and signal metadata
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)
sensors <- imotionsApi::getRespondentSensors(study, respondents[1, ])
metadata <- imotionsApi::getSensorMetaData(study, sensors[1, ])
}
}
