% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getSensors}
\alias{getSensors}
\title{Get all sensors available for a given respondent/segment.}
\usage{
getSensors(study, target, stimulus = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{target}{The target respondent/segment (an imRespondent/imSegment object as returned from
\code{\link{getRespondents}} or \code{\link{getSegments}}).}

\item{stimulus}{Optional - an imStimulus object as returned from \code{\link{getStimuli}} to retrieve sensors
specific to this stimulus. In case of a segment target, this parameter needs to be provided.}
}
\value{
An imSensorList object (data.table) with all sensors collected for the respondent of interest.
}
\description{
Retrieves detailed information about sensors available for a given respondent/segment.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)
segments <- imotionsApi::getSegments(study)
sensors <- imotionsApi::getSensors(study, respondents[1, ])

# Get sensors for a specific stimulus
stimuli <- imotionsApi::getStimuli(study)
sensors <- imotionsApi::getSensors(study, respondents[1, ], stimuli[1, ])

# Get sensors for a specific segment/Stimulus
stimuli <- imotionsApi::getStimuli(study)
sensors <- imotionsApi::getSensors(study, segments[1, ], stimuli[1, ])
}
}
