% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getSensorsUrl.imSegment}
\alias{getSensorsUrl.imSegment}
\title{getSensorsUrl method to get the path/url to the sensors of a segment object.}
\usage{
\method{getSensorsUrl}{imSegment}(study, imObject, stimulus = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent or imSegment object of interest.}

\item{stimulus}{Optional - An imStimulus object as returned from  \code{\link{getStimuli}}.}
}
\description{
getSensorsUrl method to get the path/url to the sensors of a segment object.
}
\keyword{internal}
