% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getSensorsUrl}
\alias{getSensorsUrl}
\title{Generic getSensorsUrl function that takes as parameter a study object and a respondent/segment object.}
\usage{
getSensorsUrl(study, imObject, stimulus = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent or imSegment object of interest.}

\item{stimulus}{Optional - An imStimulus object as returned from  \code{\link{getStimuli}}.}
}
\description{
Return the path/url to the sensors for this respondent/segment object.
}
\keyword{internal}
