% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getStimuli}
\alias{getStimuli}
\title{Get all stimuli from a study.}
\usage{
getStimuli(study, respondent = NULL, relevant = TRUE)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{respondent}{Optional - An imRespondent object as returned from \code{\link{getRespondents}}.}

\item{relevant}{A boolean indicating whether only relevant stimuli should be kept, by default non-relevant stimuli
are discarded.}
}
\value{
An imStimulusList object (data.table) containing all stimuli from the study.
}
\description{
Retrieves detailed information about stimuli in the study.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)

## Get all stimuli in the study
stimuli <- imotionsApi::getStimuli(study)

## Get all stimuli for a specific respondent
stimuli <- imotionsApi::getStimuli(study, respondents[1, ])
}
}
