% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getStimulus}
\alias{getStimulus}
\title{Get a specific stimulus from a study.}
\usage{
getStimulus(study, stimulusId)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{stimulusId}{The id of the stimulus you would like to retrieve.}
}
\value{
An imStimulus object (data.table) containing the stimulus of interest.
}
\description{
Available stimuli can be found with \code{\link{getStimuli}}.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
stimuli <- imotionsApi::getStimuli(study)
stimulus <- imotionsApi::getStimulus(study, stimuli$id[1])
}
}
