% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getStudiesUrl}
\alias{getStudiesUrl}
\title{Return the path/url to the studies available.}
\usage{
getStudiesUrl(connection)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}
}
\description{
Return the path/url to the studies available.
}
\keyword{internal}
