% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getStudyBaseUrl}
\alias{getStudyBaseUrl}
\title{Return the basePath/baseUrl to the iMotions server based on a specific study object.}
\usage{
getStudyBaseUrl(study)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}
}
\description{
Return the basePath/baseUrl to the iMotions server based on a specific study object.
}
\keyword{internal}
