% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getStudyS3BaseUrl}
\alias{getStudyS3BaseUrl}
\title{Return the s3basePath/s3baseUrl to the iMotions server based on a specific study object (used to download data).}
\usage{
getStudyS3BaseUrl(study)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}
}
\description{
Return the s3basePath/s3baseUrl to the iMotions server based on a specific study object (used to download data).
}
\keyword{internal}
