% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getStudySpecificUrl}
\alias{getStudySpecificUrl}
\title{Based on the kind of connection, return the path/url to a specific study object.}
\usage{
getStudySpecificUrl(study)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}
}
\description{
Based on the kind of connection, return the path/url to a specific study object.
}
\keyword{internal}
