% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getStudyUrl}
\alias{getStudyUrl}
\title{Return the path/url of a specific study object when connecting locally.}
\usage{
getStudyUrl(study)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}
}
\description{
Return the path/url of a specific study object when connecting locally.
}
\keyword{internal}
