% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getStudyUrlById}
\alias{getStudyUrlById}
\title{Return the path/url to a specific study based of its studyId.}
\usage{
getStudyUrlById(connection, studyId)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}

\item{studyId}{The id of the study of interest.}
}
\description{
Return the path/url to a specific study based of its studyId.
}
\keyword{internal}
