% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getTouchActor}
\alias{getTouchActor}
\title{Get a specific touch actor from a stimulus.}
\usage{
getTouchActor(study, stimulus, touchActorId)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{stimulus}{An imStimulus object as returned from \code{\link{getStimuli}}.}

\item{touchActorId}{The id of the touch actor you would like to retrieve.}
}
\value{
An imTouchActor object (data.table) containing the touch actor of interest, or NULL if none is defined
        for this stimulus or none matches touchActorId.
}
\description{
Available touch actors can be found with \code{\link{getTouchActors}}. Passing the result on to
\code{\link{getTouchActorAois}} (instead of the whole stimulus) scopes the underlying contact detail fetch to just
this actor - see \code{getTouchActorDetailsUrl.imTouchActor} - rather than triggering signal generation for every
touch actor on the stimulus.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
stimulus <- imotionsApi::getStimuli(study)[1, ]
touchActors <- imotionsApi::getTouchActors(study, stimulus)
touchActor <- imotionsApi::getTouchActor(study, stimulus, touchActors$id[1])
}
}
