% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getTouchActorAoi}
\alias{getTouchActorAoi}
\title{Get a specific (touch actor, AOI) combination from a stimulus.}
\usage{
getTouchActorAoi(study, stimulus, touchActorId, aoiId, respondent = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{stimulus}{An imStimulus object as returned from \code{\link{getStimuli}}.}

\item{touchActorId}{The id of the touch actor you would like to retrieve.}

\item{aoiId}{The id of the AOI you would like to retrieve.}

\item{respondent}{Optional - An imRespondent object as returned from \code{\link{getRespondents}} to attach that
respondent's contact in/out fileId/resultId to the result.}
}
\value{
An imTouchAOI object (data.table) containing the (touch actor, AOI) combination of interest, or NULL if
        touchActorId itself doesn't match, or matches but has no AOI matching aoiId.
}
\description{
Available combinations can be found with \code{\link{getTouchActorAois}}. Both a touchActorId and an aoiId are
required - unlike a gaze AOI, more than one touch actor can share the same interactive AOI, so the aoiId alone
would not uniquely identify one combination.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
stimulus <- imotionsApi::getStimuli(study)[1, ]
touchAois <- imotionsApi::getTouchActorAois(study, stimulus)
touchAoi <- imotionsApi::getTouchActorAoi(study, stimulus, touchAois$touchActorId[1], touchAois$id[1])
}
}
