% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getTouchActorsUrl}
\alias{getTouchActorsUrl}
\title{Return the path/url to the touch actors defined for a specific stimulus.}
\usage{
getTouchActorsUrl(study, stimulusId)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{stimulusId}{The id of the stimulus of interest.}
}
\description{
Return the path/url to the touch actors defined for a specific stimulus.
}
\keyword{internal}
