% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getUploadAoiMetadataUrl}
\alias{getUploadAoiMetadataUrl}
\title{getUploadAoiMetadataUrl function that takes as parameter a study object.}
\usage{
getUploadAoiMetadataUrl(study)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}
}
\description{
Return the path/url to upload AOI metadata for a specific study.
}
\keyword{internal}
