% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getUploadEventsUrl.imRespondent}
\alias{getUploadEventsUrl.imRespondent}
\title{getUploadSensorsUrl method to get the path/url to upload a sensor to this respondent object.}
\usage{
\method{getUploadEventsUrl}{imRespondent}(study, imObject)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent or imSegment object of interest.}
}
\description{
getUploadSensorsUrl method to get the path/url to upload a sensor to this respondent object.
}
\keyword{internal}
