% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getUploadEventsUrl}
\alias{getUploadEventsUrl}
\title{Generic getUploadEventsUrl function that takes as parameter a study object and a respondent/segment object.
Return the path/url to upload events to this respondent/segment object.}
\usage{
getUploadEventsUrl(study, imObject)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent or imSegment object of interest.}
}
\description{
Generic getUploadEventsUrl function that takes as parameter a study object and a respondent/segment object.
Return the path/url to upload events to this respondent/segment object.
}
\keyword{internal}
