% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getUploadMetricsUrl.imRespondent}
\alias{getUploadMetricsUrl.imRespondent}
\title{getUploadMetricsUrl method to get the path/url to upload metrics to this respondent object.}
\usage{
\method{getUploadMetricsUrl}{imRespondent}(study, imObject)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent or imSegment object of interest.}
}
\description{
getUploadMetricsUrl method to get the path/url to upload metrics to this respondent object.
}
\keyword{internal}
