% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getUploadMetricsUrl}
\alias{getUploadMetricsUrl}
\title{Generic getUploadMetricsUrl function that takes as parameter a study object and a respondent/segment object.
Return the path/url to upload metrics to this respondent/segment object.}
\usage{
getUploadMetricsUrl(study, imObject)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent or imSegment object of interest.}
}
\description{
Generic getUploadMetricsUrl function that takes as parameter a study object and a respondent/segment object.
Return the path/url to upload metrics to this respondent/segment object.
}
\keyword{internal}
