% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getUploadSensorDataUrl.imSegment}
\alias{getUploadSensorDataUrl.imSegment}
\title{getUploadSensorDataUrl method to get the path/url to upload a sensor data to this segment object.}
\usage{
\method{getUploadSensorDataUrl}{imSegment}(study, imObject, stimulus = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent or imSegment object of interest.}

\item{stimulus}{Optional - An imStimulus object as returned from  \code{\link{getStimuli}}.}
}
\description{
getUploadSensorDataUrl method to get the path/url to upload a sensor data to this segment object.
}
\keyword{internal}
