% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getUploadSensorDataUrl}
\alias{getUploadSensorDataUrl}
\title{Generic getUploadSensorDataUrl function that takes as parameter a study object and a respondent/segment object.
Return the path/url to upload a sensor data to this respondent/segment object.}
\usage{
getUploadSensorDataUrl(study, imObject, stimulus = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent or imSegment object of interest.}

\item{stimulus}{Optional - An imStimulus object as returned from  \code{\link{getStimuli}}.}
}
\description{
Generic getUploadSensorDataUrl function that takes as parameter a study object and a respondent/segment object.
Return the path/url to upload a sensor data to this respondent/segment object.
}
\keyword{internal}
