% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{imConnection}
\alias{imConnection}
\title{Create a connection with the iMotions API.}
\usage{
imConnection(token, baseUrl = NULL, s3BaseUrl = NULL, localPath = NULL)
}
\arguments{
\item{token}{The token to be used for authentication.}

\item{baseUrl}{Optional - The server to connect to in case of remote connection.}

\item{s3BaseUrl}{Optional - The server to use to write back data in case of remote connection.}

\item{localPath}{Optional - The path to write local file in case of remote connection (default to temporary folder).}
}
\value{
An imConnection object to be passed to other methods.
}
\description{
Tokens can be obtained by right-clicking on an analysis in iMotions and clicking "Get token for R API connection".
}
\examples{
\dontrun{
# Local connection
myToken <- "xxxxxxxx"
connection <- imotionsApi::imConnection(myToken)

# Remote connection
myToken <- "token"
connection <- imotionsApi::imConnection(myToken, "myBaseUrl", "myS3BaseUrl", "myLocalPath")
}
}
