% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\docType{package}
\name{imotionsApi-package}
\alias{imotionsApi}
\alias{imotionsApi-package}
\title{iMotions R API package}
\description{
A client for accessing data from the iMotions Lab or iMotions Online platform.
}
\details{
Use tokens (found on a remote study's R Analysis page or by right-clicking on an analysis in the software)
to start accessing your data.
}
\examples{
myToken <- "xxxxxxxx"
connection <- imotionsApi::imConnection(myToken)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://imotions.com}
  \item \url{https://my.imotions.com/#studies}
}

}
\author{
\strong{Maintainer}: iMotions \email{support@imotions.com}

Authors:
\itemize{
  \item Amandine Grappe
}

Other contributors:
\itemize{
  \item Paolo Masulli [contributor]
}

}
