% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{imStudy}
\alias{imStudy}
\title{Load an iMotions study by id.}
\usage{
imStudy(connection, studyId)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}

\item{studyId}{The id of the study you would like to retrieve.}
}
\value{
An imStudy object to be passed to other methods.
}
\description{
Retrieves detailed information about a study including stimuli, respondents, segments, and more.
}
\details{
Available studies can be found with \code{\link{listStudies}}.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
studyId <- studies$id[1]
study <- imotionsApi::imStudy(connection, studyId)
}
}
