% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{listLoadedStudies}
\alias{listLoadedStudies}
\title{List studies that have been loaded in the current session.}
\usage{
listLoadedStudies()
}
\value{
A list of the studies that have been loaded in the current session.
}
\description{
More detailed information about a study can be retrieved using \code{\link{imStudy}}.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
studyId <- studies$id[1]
study <- imotionsApi::imStudy(connection, studyId)
listLoadedStudies()
}
}
