% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{listStudies}
\alias{listStudies}
\title{List available studies.}
\usage{
listStudies(connection)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}
}
\value{
A data frame containing names and ids of studies.
}
\description{
More detailed information about a study can be retrieved using \code{\link{imStudy}}.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
}
}
