% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{postHttr}
\alias{postHttr}
\title{Perform a POST HTTP request with authentication.}
\usage{
postHttr(connection, url, reqBody, message = NULL)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}

\item{url}{The url/path where the POST request need to be send.}

\item{reqBody}{The body of the request.}

\item{message}{Optional - a short message indicating which steps are getting performed to get a more indicative
error message.}
}
\value{
The last HTTP response.
}
\description{
The request will be retried up to 3 times if an error is encountered.
}
\keyword{internal}
