% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{postJSON}
\alias{postJSON}
\title{Post a JSON file to the indicated path/url - optionally allow the user to add more information about to know
when an error occurred.}
\usage{
postJSON(connection, url, postData, message = NULL)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}

\item{url}{The url/path where the JSON file will be located.}

\item{postData}{The JSON body that needs to be uploaded.}

\item{message}{Optional - a short message indicating which steps are getting performed to get a more indicative
error message.}
}
\value{
The url/path where the JSON file has been uploaded and additional information.
}
\description{
Post a JSON file to the indicated path/url - optionally allow the user to add more information about to know
when an error occurred.
}
\keyword{internal}
