% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateAoiFiltering}
\alias{privateAoiFiltering}
\title{Private generic function to filter the list of AOIs to return based on stimulus/respondent object.}
\usage{
privateAoiFiltering(study, stimulus = NULL, respondent = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{stimulus}{Optional - An imStimulus object as returned from \code{\link{getStimuli}}.}

\item{respondent}{Optional - An imRespondent object as returned from \code{\link{getRespondents}}.}
}
\value{
A data.table with AOIs filtered.
}
\description{
Private generic function to filter the list of AOIs to return based on stimulus/respondent object.
}
\keyword{internal}
