% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateCalculateAreaAoi}
\alias{privateCalculateAreaAoi}
\title{Private function to compute the area of an AOI (adapted from pracma::polyarea).}
\usage{
privateCalculateAreaAoi(x, y)
}
\arguments{
\item{x}{The X coordinates of the polygon.}

\item{y}{The Y coordinates of the polygon.}
}
\value{
A data.table with AOIs.
}
\description{
Private function to compute the area of an AOI (adapted from pracma::polyarea).
}
\keyword{internal}
