% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateCreateDataHeader}
\alias{privateCreateDataHeader}
\title{Create data header for the file that needs to be uploaded.}
\usage{
privateCreateDataHeader(params, sensorName, scriptName, fileDependency = NULL)
}
\arguments{
\item{params}{The list of parameters provided to the script - specific parameters value will be stored as metadata.}

\item{sensorName}{The name of the new sensor you would like to create.}

\item{scriptName}{The name of the script used to produce these signals/metrics.}
}
\description{
Create data header for the file that needs to be uploaded.
}
\keyword{internal}
