% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateCreateHeader}
\alias{privateCreateHeader}
\title{Create header for the file that needs to be uploaded.}
\usage{
privateCreateHeader(params, data, sensorName, scriptName)
}
\arguments{
\item{params}{The list of parameters provided to the script - specific parameters value will be stored as metadata.}

\item{data}{A data.table containing the signals to upload (imData object are also accepted).}

\item{sensorName}{The name of the new sensor you would like to create.}

\item{scriptName}{The name of the script used to produce these signals.}
}
\description{
Create header for the file that needs to be uploaded.
}
\keyword{internal}
