% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateCreatePostRequest}
\alias{privateCreatePostRequest}
\title{Create headers specific to the data format (signals, events, exports).}
\usage{
privateCreatePostRequest(
  params,
  study,
  sampleName,
  fileName,
  overwrite = TRUE,
  data_names = NULL
)
}
\arguments{
\item{params}{The list of parameters provided to the script - specific parameters value will be stored as metadata.}

\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{sampleName}{The name of the new sensor/metrics/events/export you would like to create. In case of export,
this should correspond to the name of the file created (a path can also be provided).}

\item{overwrite}{Optional - a boolean indicating if the new sensor/metrics/events/export should overwrite
sensor/metrics/events/export generated for the same flowName. By default, they are overwritten.}

\item{data_names}{Optional names of the data columns included in an online sensor upload.}
}
\description{
Create headers specific to the data format (signals, events, exports).
}
\keyword{internal}
