% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateDownloadData}
\alias{privateDownloadData}
\title{Return metrics / signals (with modified Timestamp if needed) for the sensor of interest.}
\usage{
privateDownloadData(study, sensor, signalsName = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{sensor}{An imSensor object as returned from \code{\link{getRespondentSensors}}.}

\item{signalsName}{Optional - A vector of specific signals name you would like to return.}
}
\value{
A data.table with all signals (or specified signals) from the sensor of interest.
}
\description{
Return metrics / signals (with modified Timestamp if needed) for the sensor of interest.
}
\keyword{internal}
