% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateGetAoiRespondentMetricsPath}
\alias{privateGetAoiRespondentMetricsPath}
\title{Return the local cache path for AOI respondent metrics.}
\usage{
privateGetAoiRespondentMetricsPath(study, AOI, respondent)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{AOI}{An imAOI object as returned from \code{\link{getAois}}.}

\item{respondent}{An imRespondent object as returned from \code{\link{getRespondents}}.}
}
\description{
Return the local cache path for AOI respondent metrics.
}
\keyword{internal}
