% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateGetFileHeader}
\alias{privateGetFileHeader}
\title{Create headers specific to the data format (signals, events, exports).}
\usage{
privateGetFileHeader(
  data,
  params = NULL,
  sampleName = NULL,
  scriptName = NULL,
  metadata = NULL
)
}
\arguments{
\item{data}{A data.table containing the signals/metrics/events to upload (imData object are also accepted).}

\item{params}{The list of parameters provided to the script - specific parameters value will be stored as metadata.}

\item{sampleName}{The name of the new sensor/metrics/events you would like to create.}

\item{scriptName}{The name of the script used to produce these signals.}

\item{metadata}{Optional - a data.table with metadata information. Column names will be converted to metadata headers
and there must be a row corresponding to each data column.}
}
\description{
Create headers specific to the data format (signals, events, exports).
}
\keyword{internal}
