% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateGetIntervalsForScenes}
\alias{privateGetIntervalsForScenes}
\title{Get time intervals for scenes.}
\usage{
privateGetIntervalsForScenes(study, respondent)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{respondent}{An imRespondent object as returned from \code{\link{getRespondents}}.}
}
\value{
A data.table composed of the start, end, duration, parent stimulus, id and name of each scene.
}
\description{
Return the start, end, duration, parent stimulus, id and name of each scene.
Scene with multiple fragments will be return as multiple intervals (start, end, duration) with same id and name.
}
\keyword{internal}
