% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateRespondentFiltering.imAOI}
\alias{privateRespondentFiltering.imAOI}
\title{Get respondents for a specific AOI.}
\usage{
\method{privateRespondentFiltering}{imAOI}(study, obj)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{obj}{An imAOI object as returned from \code{\link{getAois}} for which you would like to get the respondents.}
}
\value{
A data.table with all respondents for whom a specific AOI has been defined.
}
\description{
S3 method to get all respondents for a specific AOI.
}
\keyword{internal}
