% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateRespondentFiltering.imStimulus}
\alias{privateRespondentFiltering.imStimulus}
\title{Get respondents for a specific stimulus.}
\usage{
\method{privateRespondentFiltering}{imStimulus}(study, obj)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{obj}{An imStimulus object as returned from \code{\link{getStimuli}} for which you would like to get
the respondents.}
}
\value{
A data.table with all respondents exposed to a specific stimulus.
}
\description{
S3 method to get all respondents for a specific stimulus.
}
\keyword{internal}
