% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateRespondentFiltering.imStudy}
\alias{privateRespondentFiltering.imStudy}
\title{Get all respondents from a study.}
\usage{
\method{privateRespondentFiltering}{imStudy}(study, ...)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{...}{Further arguments passed from the \code{\link{privateRespondentFiltering}} method.}
}
\value{
A data.table with all respondents for the study.
}
\description{
S3 method default method to get all respondents for a study.
}
\keyword{internal}
