% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateRespondentFiltering}
\alias{privateRespondentFiltering}
\title{Private generic function to filter the list of respondents to return based on stimulus/AOI object.}
\usage{
privateRespondentFiltering(study, obj = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{obj}{Optional - An imAOI or imStimulus object to filter respondents.}
}
\value{
A data.table with respondents filtered.
}
\description{
Private generic function to filter the list of respondents to return based on stimulus/AOI object.
}
\keyword{internal}
