% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateSaveToFile}
\alias{privateSaveToFile}
\title{Create a temporary file with the data/metadata that needs to be uploaded.}
\usage{
privateSaveToFile(
  params,
  data,
  sensorName = NULL,
  scriptName = NULL,
  metadata = NULL
)
}
\arguments{
\item{params}{The list of parameters provided to the script - specific parameters value will be stored as metadata.}

\item{data}{A data.table containing the signals to upload (imData object are also accepted).}

\item{sensorName}{The name of the new sensor you would like to create.}

\item{scriptName}{The name of the script used to produce these signals.}

\item{metadata}{Optional - a data.table with metadata information. Column names will be converted to metadata headers
and there must be a row corresponding to each data column.}
}
\description{
Create a temporary file with the data/metadata that needs to be uploaded.
}
\keyword{internal}
