% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateUpload}
\alias{privateUpload}
\title{Upload signals/events to a given respondent/segment.}
\usage{
privateUpload(
  params,
  study,
  data,
  target,
  sensorName = NULL,
  scriptName = NULL,
  metadata = NULL,
  stimulus = NULL
)
}
\arguments{
\item{params}{The list of parameters provided to the script - specific parameters value will be stored as metadata.}

\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{data}{A data.table containing the signals to upload (imData object are also accepted).}

\item{target}{The target respondent for the sensor (an imRespondent object as returned from
\code{\link{getRespondents}}).}

\item{sensorName}{The name of the new sensor you would like to create.}

\item{scriptName}{The name of the script used to produce these signals.}

\item{metadata}{Optional - a data.table with metadata information. Column names will be converted to metadata headers
and there must be a row corresponding to each data column.}

\item{stimulus}{Optional - an imStimulus object as returned from \code{\link{getStimuli}} to upload data specific to
this stimulus.}
}
\description{
Upload signals/events to a given respondent/segment.
}
\keyword{internal}
