% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateUploadAoiMetrics.imRespondent}
\alias{privateUploadAoiMetrics.imRespondent}
\title{Write back or upload metrics for a respondent of interest.}
\usage{
\method{privateUploadAoiMetrics}{imRespondent}(study, obj, AOI, metrics)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{obj}{An imRespondent or imSegment object of interest.}

\item{AOI}{An imAOI object as returned from \code{\link{getAois}}, or an imTouchAOI object as returned from
\code{\link{getTouchActorAois}}.}

\item{metrics}{A data.table containing the metrics to upload.}
}
\description{
S3 method to write back metrics locally or upload them for a specific respondent.
}
\keyword{internal}
