% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateUploadAoiMetrics.imSegment}
\alias{privateUploadAoiMetrics.imSegment}
\title{Upload metrics for a segment of interest.}
\usage{
\method{privateUploadAoiMetrics}{imSegment}(study, obj, AOI, metrics)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{obj}{An imRespondent or imSegment object of interest.}

\item{AOI}{An imAOI object as returned from \code{\link{getAois}}.}

\item{metrics}{A data.table containing the metrics to upload.}
}
\description{
S3 method to upload metrics for a specific segment.
}
\keyword{internal}
