% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateUploadAoiMetrics}
\alias{privateUploadAoiMetrics}
\title{Generic privateUploadAoiMetrics function that takes as parameter a study object, a respondent/segment object, an AOI
and some metrics to upload.}
\usage{
privateUploadAoiMetrics(study, obj, AOI, metrics)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{obj}{An imRespondent or imSegment object of interest.}

\item{AOI}{An imAOI object as returned from \code{\link{getAois}}.}

\item{metrics}{A data.table containing the metrics to upload.}
}
\description{
Writes back metrics locally or upload them based on the workflow selected.
}
\keyword{internal}
