% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateUploadToPresignedUrl}
\alias{privateUploadToPresignedUrl}
\title{Get upload credentials and upload a file to the returned presigned URL.}
\usage{
privateUploadToPresignedUrl(
  connection,
  uploadUrl,
  postData,
  fileName,
  credentialsMsg,
  uploadMsg
)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}

\item{uploadUrl}{The URL used to request upload credentials.}

\item{postData}{The JSON body used to request upload credentials.}

\item{fileName}{The path of the file to upload.}

\item{credentialsMsg}{Message used if requesting credentials fails.}

\item{uploadMsg}{Message used if uploading the file fails.}
}
\value{
The upload credentials returned by the API.
}
\description{
Get upload credentials and upload a file to the returned presigned URL.
}
\keyword{internal}
