% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{reorderColnames}
\alias{reorderColnames}
\title{Reorder columns of a data.table based on a vector of column names.}
\usage{
reorderColnames(data, explicitlyOrdered)
}
\arguments{
\item{data}{A data.table with columns to reorder.}

\item{explicitlyOrdered}{An vector of column names explicitly ordered.}
}
\value{
A data.table with reordered columns.
}
\description{
Reorder columns of a data.table based on a vector of column names.
}
\keyword{internal}
