% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{reorderSensorColumns}
\alias{reorderSensorColumns}
\title{Make sensors columns order a bit more intuitive, and reliable.}
\usage{
reorderSensorColumns(sensors)
}
\arguments{
\item{sensors}{A sensors data.table to reoder.}
}
\value{
A sensors data.table with reordered columns.
}
\description{
Make sensors columns order a bit more intuitive, and reliable.
}
\keyword{internal}
