% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{retryHttr}
\alias{retryHttr}
\title{Perform a GET, POST or PUT HTTP request with authentication. Retries up to 3 times if an error is encountered.}
\usage{
retryHttr(message, ...)
}
\arguments{
\item{message}{A short message indicating which steps are getting performed to get a more indicative error message.}

\item{...}{Further arguments passed from the getHttr/postHttr/putHttr method.}
}
\value{
A valid HTTP response.
}
\description{
Perform a GET, POST or PUT HTTP request with authentication. Retries up to 3 times if an error is encountered.
}
\keyword{internal}
