% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{stopOnHttpError}
\alias{stopOnHttpError}
\title{Analyse a status code from an HTTP response and stop if an error was encountered.}
\usage{
stopOnHttpError(response, message = NULL)
}
\arguments{
\item{response}{An HTTP response.}

\item{message}{A short message indicating which steps are getting performed to get a more indicative error message.}
}
\description{
Analyse a status code from an HTTP response and stop if an error was encountered.
}
\keyword{internal}
