% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{tokenHeaders}
\alias{tokenHeaders}
\title{Return token authentication header}
\usage{
tokenHeaders(token)
}
\arguments{
\item{token}{The token to be used for authentication.}
}
\description{
Return token authentication header
}
\keyword{internal}
