% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{unloadStudies}
\alias{unloadStudies}
\title{Remove studies that have been loaded in the current session.}
\usage{
unloadStudies()
}
\value{
"Studies successfully removed from the current session" if success.
}
\description{
This function can be used when new changes have been made to already loaded studies.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
studyId <- studies$id[1]
study <- imotionsApi::imStudy(connection, studyId)
unloadStudies()
}
}
