% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{checkDataFormat}
\alias{checkDataFormat}
\title{Check the format of a data.table and classify it as an imMetrics, imSignal or, imExport object.}
\usage{
checkDataFormat(data)
}
\arguments{
\item{data}{A data.table containing the signals/metrics.}
}
\value{
An imSignals, imMetrics or imExport object if data was well formatted.
}
\description{
Check the format of a data.table and classify it as an imMetrics, imSignal or, imExport object.
}
\keyword{internal}
