% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{formatGender}
\alias{formatGender}
\title{Reformat the gender column.}
\usage{
formatGender(gender)
}
\arguments{
\item{gender}{The gender column to reformat.}
}
\value{
The gender column coded as Male/Female/Other
}
\description{
Converts 0/1/4 values into MALE/FEMALE/OTHER values.
}
\keyword{internal}
