% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAOI}
\alias{getAOI}
\title{Get a specific AOI from a study.}
\usage{
getAOI(study, AOIId)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{AOIId}{The id of the AOI you would like to retrieve.}
}
\value{
An imAOI object (data.table) containing the AOI of interest.
}
\description{
Available AOIs can be found with \code{\link{getAOIs}}. In case no AOI is found, return NULL.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
AOIs <- imotionsApi::getAOIs(study)
AOI <- imotionsApi::getAOI(study, AOIs$id[1])
}
}
