% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAOIDetailsUrl}
\alias{getAOIDetailsUrl}
\title{Generic getAOIDetailsUrl function that takes as parameter a study object, an AOI/stimulus object and optionally
a respondent object.}
\usage{
getAOIDetailsUrl(study, imObject, respondent = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imAOI or imStimulus object of interest.}

\item{respondent}{Optional - An imRespondent object as returned from  \code{\link{getRespondents}}.}
}
\description{
Return the path/url to a specific AOI's details information or to all AOIs' details for a specific stimulus.
}
\keyword{internal}
