% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getAOIRespondentMetrics}
\alias{getAOIRespondentMetrics}
\title{Get the metrics for a specific AOI/respondent combination.}
\usage{
getAOIRespondentMetrics(study, AOI, respondent)
}
\arguments{
\item{study}{An imStudy object as returned from imStudy()}

\item{AOI}{An imAOI object as returned from \code{\link{getAOIs}}.}

\item{respondent}{An imRespondent object as returned from \code{\link{getRespondents}}.}
}
\value{
A data.table of one row (imMetrics object) with metrics for the AOI /respondent combination of interest.
}
\description{
Get the metrics for a specific AOI/respondent combination.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
AOIs <- imotionsApi::getAOIs(study)
respondents <- imotionsApi::getRespondents(study, AOI = AOIs[1, ])
AOImetrics <- imotionsApi::getAOIRespondentMetrics(study, AOIs[1, ], respondents[1, ])
}
}
