% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getFile}
\alias{getFile}
\title{Retrieve a zip/csv file from the indicated path/url - optionally allow the user to add more information to know
when an error occurred.}
\usage{
getFile(connection, url, message = NULL, fileName = NULL)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}

\item{url}{The url/path where the zip/csv file is located.}

\item{message}{Optional - a short message indicating which steps are getting performed to get a more indicative
error message.}

\item{fileName}{Optional - In case of a zip folder, the name of the file to look for.}
}
\value{
A list with the temporary folder path and the downloaded file name.
}
\description{
Retrieve a zip/csv file from the indicated path/url - optionally allow the user to add more information to know
when an error occurred.
}
\keyword{internal}
