% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getHttr}
\alias{getHttr}
\title{Perform a GET HTTP request with authentication.}
\usage{
getHttr(connection, url, message = NULL, writePath = NULL)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}

\item{url}{The url/path where the file is located.}

\item{message}{Optional - a short message indicating which steps are getting performed to get a more indicative
error message.}

\item{writePath}{Optional - A local path where the response body should be streamed.}
}
\value{
The last response.
}
\description{
The request will be retried up to 3 times if an error is encountered.
}
\keyword{internal}
