% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getHttr}
\alias{getHttr}
\title{Perform a GET HTTP request with authentication.}
\usage{
getHttr(connection, url)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}

\item{url}{The url/path where the JSON file is located.}
}
\value{
The last response.
}
\description{
The request will be retried up to 3 times if an error is encountered.
}
\keyword{internal}
