% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{getUploadAoiMetricsUrl.imSegment}
\alias{getUploadAoiMetricsUrl.imSegment}
\title{getUploadAoiMetricsUrl method to get the path/url to upload AOI metrics to this segment object.}
\usage{
\method{getUploadAoiMetricsUrl}{imSegment}(study, imObject, AOI)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imRespondent or imSegment object of interest.}

\item{AOI}{An imAOI object as returned from \code{\link{getAOIs}}.}
}
\description{
getUploadAoiMetricsUrl method to get the path/url to upload AOI metrics to this segment object.
}
\keyword{internal}
