% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{imConnection}
\alias{imConnection}
\title{Create a connection with the iMotions API.}
\usage{
imConnection(token, baseUrl = "https://my.imotions.com/api")
}
\arguments{
\item{token}{The token to be used for authentication.}

\item{baseUrl}{The iMotions server to connect to, normally you do not need to change the default.}
}
\value{
An imConnection object to be passed to other methods.
}
\description{
Tokens can be found on a study's R Analysis page or given directly by the iMotions software
}
\examples{
\dontrun{
myToken <- "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
connection <- imotionsApi::imConnection(myToken)
}
}
