% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\docType{package}
\name{imotionsApi-package}
\alias{imotionsApi}
\alias{imotionsApi-package}
\title{iMotions R API package}
\description{
A client for accessing data from the iMotions Biometrics Research Platform.
}
\details{
Use tokens (found on a study's R Analysis page) to load information from the
iMotions API
}
\examples{
myToken <- "xxxxx-xxxx-xxxxx-xxxxx"
connection <- imotionsApi::imConnection(myToken)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://imotions.com}
  \item \url{https://my.imotions.com/#studies}
}

}
