% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateAOIFiltering.imRespondent}
\alias{privateAOIFiltering.imRespondent}
\title{Get AOIs for a specific respondent.}
\usage{
\method{privateAOIFiltering}{imRespondent}(study, stimulus = NULL, respondent)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{stimulus}{Optional - An imStimulus object as returned from  \code{\link{getStimuli}}.}

\item{respondent}{An imRespondent object as returned from \code{\link{getRespondents}} for which you want to get the
AOIs.}
}
\value{
A data.table with all AOIs defined for a specific respondent.
}
\description{
S3 method to get all AOIs for a specific respondent.
}
