% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateAOIFiltering.imStimulus}
\alias{privateAOIFiltering.imStimulus}
\title{Get AOIs for a specific stimulus.}
\usage{
\method{privateAOIFiltering}{imStimulus}(study, stimulus, ...)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{stimulus}{An imStimulus object as returned from \code{\link{getStimuli}} for which you want to get the AOIs.}

\item{...}{Further arguments passed from the \code{\link{privateAOIFiltering}} method.}
}
\value{
A data.table with all AOIs defined for a specific stimulus.
}
\description{
S3 method to get all AOIs for a specific stimulus.
}
\keyword{internal}
