% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateAOIFiltering.imStudy}
\alias{privateAOIFiltering.imStudy}
\title{Get all AOIs from a study.}
\usage{
\method{privateAOIFiltering}{imStudy}(study, ...)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{...}{Further arguments passed from the \code{\link{privateAOIFiltering}} method.}
}
\value{
A data.table with all AOIs for the study.
}
\description{
S3 method default method to get all AOIs for a study.
}
\keyword{internal}
