% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateCreateHeader}
\alias{privateCreateHeader}
\title{Create header for the file that needs to be uploaded.}
\usage{
privateCreateHeader(params, data, sampleName, scriptName)
}
\arguments{
\item{params}{The list of parameters provided to the script - specific parameters value will be stored as metadata.}

\item{data}{A data.table containing the signals/metrics/events to upload (imData object are also accepted).}

\item{sampleName}{The name of the new sensor/metrics/events/export you would like to create. In case of export,
this should correspond to the name of the file created (a path can also be provided).}

\item{scriptName}{The name of the script used to produce these signals.}
}
\description{
Create header for the file that needs to be uploaded.
}
\keyword{internal}
