% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{privateGetAOIDetails}
\alias{privateGetAOIDetails}
\title{Private function that takes as parameter a study object, an imObject(imAOI or imStimulus) object and, optionally
an imRespondent object.}
\usage{
privateGetAOIDetails(study, imObject, respondent = NULL)
}
\arguments{
\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{imObject}{An imAOI or imStimulus object as returned from \code{\link{getAOIs}} or \code{\link{getStimuli}}.
For remote study, only an imAOI object should ever be used.}

\item{respondent}{Optional - An imRespondent object as returned from \code{\link{getRespondents}}.}
}
\value{
A data.frame with details about the AOI of interest (i.e. respondents for whom it was defined and file paths
        to their IsGazeInAOI data).
}
\description{
Return details for a specific AOI or all AOIs defined for a stimulus (i.e. respondents for whom it was defined and
file paths to their corresponding IsGazeInAOI data).
}
\keyword{internal}
