% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{putHttr}
\alias{putHttr}
\title{Perform a PUT HTTP request with authentication.}
\usage{
putHttr(connection, url, reqBody)
}
\arguments{
\item{connection}{An imConnection object as returned from \code{\link{imConnection}}.}

\item{url}{The url/path where the JSON file is located.}

\item{reqBody}{The body of the request.}
}
\value{
The last HTTP response.
}
\description{
The request will be retried up to 3 times if an error is encountered.
}
\keyword{internal}
