% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{truncateSignalsByIntervals}
\alias{truncateSignalsByIntervals}
\title{Truncate signals data based on given intervals.}
\usage{
truncateSignalsByIntervals(signals, intervals, dropIntervals = FALSE)
}
\arguments{
\item{signals}{An imSignals object as returned by \code{\link{getSensorData}} or a data.table including a
"Timestamp" column that needs to be cut.}

\item{intervals}{An imInterval or imIntervalList object with start/end of data to subset as given by
\code{\link{getRespondentIntervals}} or \code{\link{getAOIRespondentData}}.}

\item{dropIntervals}{A boolean indicating whether Timestamps falling between an interval start/end should be
discarded, by default there are kept and other Timestamps are removed.}
}
\value{
A truncated imSignals object.
}
\description{
Any interval combination can be asked. Timestamps falling between an interval start/end will be kept, others will be
discarded. If dropIntervals is set to TRUE, Timestamps falling between an interval start/end will be discarded.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)
sensors <- imotionsApi::getSensors(study, respondents[1, ])
signals <- imotionsApi::getSensorData(study, sensors[3, ])
intervals <- imotionsApi::getRespondentIntervals(study, respondents[1, ])

# get the 3 first intervals
dataSubset <- imotionsApi::truncateSignalsByIntervals(signals, intervals[1:3, ])

# remove the second interval
dataSubset <- imotionsApi::truncateSignalsByIntervals(signals, intervals[2, ], dropIntervals = TRUE)
}
}
