% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{uploadAOIMetadata}
\alias{uploadAOIMetadata}
\title{Upload AOI metrics metadata for a specific study. The "Group" and "Group description" fields are mandatory
as they are used to group the metadata.}
\usage{
uploadAOIMetadata(study, metadata)
}
\arguments{
\item{study}{An imStudy object as returned from imStudy()}

\item{metadata}{A data.table with metadata information. Column names will be converted to metadata headers.}
}
\description{
Upload AOI metrics metadata for a specific study. The "Group" and "Group description" fields are mandatory
as they are used to group the metadata.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
metadata <- data.table("Units" = c("ms", ""), "Show" = c("FALSE", "TRUE"), "Group" = c("1", "2"),
                       "Group description" = c("1", "2"))

uploadAOIMetadata(study, metadata)
}
}
