% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{uploadAOIMetrics}
\alias{uploadAOIMetrics}
\title{Upload metrics for a specific respondent and AOI in a study.}
\usage{
uploadAOIMetrics(study, AOI, target, metrics)
}
\arguments{
\item{study}{An imStudy object as returned from imStudy()}

\item{AOI}{An imAOI object as returned from \code{\link{getAOIs}}.}

\item{target}{The target respondent/segment for the sensor (an imRespondent/imSegment object as returned from
\code{\link{getRespondents}} or \code{\link{getSegments}}).}

\item{metrics}{A data.table containing the metrics to upload.}
}
\description{
Upload metrics for a specific respondent and AOI in a study.
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
AOIs <- imotionsApi::getAOIs(study)
respondents <- imotionsApi::getRespondents(study, AOI = AOIs[1, ])
metrics <- data.frame("metric1" = 2, "metric2" = 234, "metric3" = 1234)
uploadAOIMetrics(study, AOIs[1, ], respondents[1, ], metrics)
}
}
