% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imotionsApi.R
\name{uploadEvents}
\alias{uploadEvents}
\title{Create events for a specific respondent in a study.}
\usage{
uploadEvents(params, study, data, target, metadata = NULL)
}
\arguments{
\item{params}{The list of parameters provided to the script - specific parameters value will be stored as metadata.}

\item{study}{An imStudy object as returned from \code{\link{imStudy}}.}

\item{data}{A data.table containing the signals/metrics to upload (imData object are also accepted).}

\item{target}{The target respondent for the sensor (an imRespondent object as returned from
\code{\link{getRespondents}}).}

\item{metadata}{Optional - a data.table with metadata information. Column names will be converted to metadata headers
and there must be a row corresponding to each data column.}
}
\description{
Events data.table must be composed of a EventName, Timestamp and Description column. Description will be rendered
as tooltip in the software.
}
\details{
Params required field are "iMotionsVersion" and "flowName" (flow name will be used to link events to the original
script)
}
\examples{
\dontrun{
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)

data <- data.table("Timestamp" = seq(1:5), "EventName" = rep("My event names", 5),
                   "Description" = rep("Description of event", 5))

params <- list("iMotionsVersion" = 8, "flowName" = "Test")
uploadEvents(params, study, data, respondents[1, ])


# Adding some metadata to the data
metadata <- data.table("Units" = c("ms", "", ""), "Show" = c("FALSE", "TRUE", "TRUE"))
uploadEvents(params, study, data, respondents[1, ], metadata)
}
}
